Fix JS export Long compatibility and update tests with real API data
- Replace Long with Int/Double in @JsExport classes to fix JS build
  (PostStamp.count, AuthObtainTokenResponse.expiresIn, dataSize)
- Update test IDs with real post/user UUIDs
- Add MIXI2_STREAM_HOST and MIXI2_TEST_ROOM_ID to secrets config
- Add EventStreamTest for server-streaming gRPC verification
- Use TEST_ROOM_ID from secrets in ChatTest
- Remove JavaCompile release(11) from stream module to fix JVM 17 dep

+++ core/src/commonMain/kotlin/work/socialhub/kmixi2/entity/PostStamp.kt
@@ -11,5 +11,5 @@ class PostStamp {
     var stamp: MediaStamp? = null
 
     @SerialName("count")
-    var count: Long = 0
+    var count: Int = 0
 }
